extract page into a new function
